Adds Main Menu
The modification in the branch is the addition of a barebones Main Menu scene with a settings screen and a start button that will load our other Test Scene. None of the toggles/sliders in the Options menu are plugged in to anything yet, but I think this addition will add a good starting point for us to build upon.